Project Page Done
Finished working on the project page added some CSS for this page for some neat animation when the project is selected also can show more information about the project itself. Also added some CSS for the nav bar to make it a little more nicer! Converted most of the imgs to webp as well.